requirements.txt
Flask==3.0.3
requests==2.32.3
google-generativeai==0.8.5
gunicorn==22.0.0